Update patch set 1
Patch Set 1:

Just an FYI, with this change, in the Play Store, m7 devices will now show as "HTC HTC One" rather than just "HTC One". Apps that just call it "One", are simply pulling only one prop, which is wrong. In my opinion, we should not do workarounds for poor app development. Almost no other devices include the manufacturer in the product name, and it is not the intention of that prop for it to include the product name.

Patch-set: 1
Reviewer: Kyle Laker <3803@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0